perf: move 40 browser-free tests from Playwright to vitest (#672)
Tests that don't interact with webviews now run in vitest without
a browser, cutting per-test time from 4-6s to ~750ms.

Moved to tests/unit/:
- list-files (11 tests), list-tests (14 tests), source-map (8 tests)
- decorations (2), global-errors (2), problems (2), run-annotations (1)

Remaining in tests/playwright/:
- Tests using enableConfigs/enableProjects (need webview settings panel)
- list-files (3), list-tests (6) — webview-dependent subset

Infrastructure:
- mock-activate.ts — vitest activate() helper with null browser
- expect-helpers.ts — polling matchers (expectTestTree, expectConnectionLog)
- VSCode mock accepts null browser (skips webview page creation)

Results: 113 vitest (35s) + 112 Playwright (3.1min) = 225 total
