+++ packages/vscode/tests/unit/run-annotations.test.ts
@@ -14,12 +14,15 @@
  * limitations under the License.
  */
 
-import { expect, test } from './utils';
+import { describe, it, expect } from 'vitest';
+import { activate } from './mock-activate';
+import { expectConnectionLog } from './expect-helpers';
 
-test('should mark test as skipped', async ({ activate }) => {
-  const { vscode, testController } = await activate({
-    'playwright.config.js': `module.exports = { testDir: 'tests' }`,
-    'tests/test.spec.ts': `
+describe('run annotations', () => {
+  it('should mark test as skipped', async () => {
+    const { vscode, testController } = await activate({
+      'playwright.config.js': `module.exports = { testDir: 'tests' }`,
+      'tests/test.spec.ts': `
       import { test } from '@playwright/test';
       test('pass', async () => {});
       test('skipped', async () => {
@@ -33,10 +36,10 @@ test('should mark test as skipped', async ({ activate }) => {
         expect(1).toBe(2);
       });
     `,
-  });
+    });
 
-  const testRun = await testController.run();
-  expect(testRun.renderLog()).toBe(`
+    const testRun = await testController.run();
+    expect(testRun.renderLog()).toBe(`
     tests > test.spec.ts > pass [2:0]
       enqueued
       started
@@ -55,15 +58,16 @@ test('should mark test as skipped', async ({ activate }) => {
       passed
   `);
 
-  await expect(vscode).toHaveConnectionLog([
-    { method: 'listFiles', params: {} },
-    { method: 'runGlobalSetup', params: {} },
-    {
-      method: 'runTests',
-      params: expect.objectContaining({
-        locations: [],
-        testIds: undefined
-      })
-    },
-  ]);
+    await expectConnectionLog(vscode, [
+      { method: 'listFiles', params: {} },
+      { method: 'runGlobalSetup', params: {} },
+      {
+        method: 'runTests',
+        params: expect.objectContaining({
+          locations: [],
+          testIds: undefined
+        })
+      },
+    ]);
+  });
 });
